Strict Whitelisting by Environment
Implements environment configuration that augments whitelisting behavior for development and production.

In development, StrongArms will unconditionally raise exceptions if it receives key value pairs it does not expect. This is to aid continuous development efforts where incoming request data matches backend system expectations; ensuring client and server to be in sync.

In production, StrongArms will log unexpected parameters to STDERR instead of raising an exception and returning 500 status. The expectation here is that your StrongArms have been fine tuned in development and logging is there for catching edge cases that may or may not affect backed behavior.

Not included in this PR -- but planned for future work -- is an explicit strict mode for production, called on an individual StrongArms basis. The explicit strict mode (production only) would omit unexpected parameters in addition to logging the event; effectively cleaning incoming params.

The behavior described above could very well become a default in production instead of a per StrongArm option.